+++ src/services/fileStorage/strategies/interface.js
@@ -9,8 +9,12 @@ class AbstractFileStorageStrategy {
 		throw new TypeError('connect method has to be implemented.');
 	}
 
+	getBucket() {
+		throw new TypeError('getBucket method has to be implemented');
+	}
+
 	listBucketsNames() {
-		throw new TypeError('list buckets method has to be implemented');
+		throw new TypeError('listBucketsNames method has to be implemented');
 	}
 
 	create() {